Move StatisticsCell logic from ParticipatoryProcesses to Core
