+++ decidim-core/app/cells/decidim/statistic/show.erb
