+++ decidim-participatory_processes/app/views/decidim/participatory_processes/participatory_processes/show.html.erb
@@ -143,7 +143,7 @@
   </div>
 
   <% if current_participatory_space.show_statistics? %>
-    <%= cell "decidim/participatory_processes/statistics", stats.collection %>
+    <%= cell "decidim/statistics", stats.collection %>
   <% end %>
 
   <% if current_participatory_space.show_metrics? %>
